Added graphic point attribute API for setting label field, font, and glyph. Added C++ font API and fixed a few C++ API errors. Issue 3361.
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/zinc/library/trunk@9648 3c1deb5b-d424-0410-962d-aba41a686d42

+++ core/source/api/zinc/graphicsmaterial.hpp
@@ -119,7 +119,7 @@ public:
 			static_cast<Cmiss_graphics_material_attribute>(attribute), value);
 	}
 
-	int getAttributeReal(Attribute attribute)
+	double getAttributeReal(Attribute attribute)
 	{
 		return Cmiss_graphics_material_get_attribute_real(id,
 			static_cast<Cmiss_graphics_material_attribute>(attribute));
